Add ApiResponseService to OpenAIApiClient for debugging
A new feature has been added to the "OpenAIApiClient" class for debugging purposes. An instance of "ApiResponseService" is now injected into the constructor of "OpenAIApiClient" and used to save responses from the OpenAI API. This will allow us to keep a record of the responses that we get from the API, which will be very useful when debugging any issues. The feature is only enabled for non-prod environments.